+++ ios/Classes/FlutterSentryPlugin.swift
@@ -3,7 +3,7 @@ import UIKit
 import Sentry
 
 public class FlutterSentryPlugin: NSObject, FlutterPlugin {
-    private static let SentryDSNKey = "SentryDSN"
+    private static let SentryDSN = "SentryDSN"
     private static let SentryEnableAutoSessionTracking =
         "SentryEnableAutoSessionTracking"
     private static let SentrySessionTrackingIntervalMillis =
@@ -18,14 +18,16 @@ public class FlutterSentryPlugin: NSObject, FlutterPlugin {
     }
 
     override init() {
+        var options = [String: Any]()
+
         guard let dsn = Bundle.main.object(
-            forInfoDictionaryKey: FlutterSentryPlugin.SentryDSNKey
+            forInfoDictionaryKey: FlutterSentryPlugin.SentryDSN
         ) else {
             NSException.raise(
                 NSExceptionName.invalidArgumentException,
                 format:"The value for key %@ is not set in Info.plist",
                 arguments: getVaList([
-                    Self.SentryDSNKey,
+                    Self.SentryDSN,
                 ]))
             return
         }
@@ -35,38 +37,48 @@ public class FlutterSentryPlugin: NSObject, FlutterPlugin {
                 NSExceptionName.invalidArgumentException,
                 format:"The value for key %@ is not a <string> type: %@",
                 arguments: getVaList([
-                    Self.SentryDSNKey,
+                    Self.SentryDSN,
                     String(describing: dsn),
                 ]))
             return
         }
+        options["dsn"] = dsnString
 
-        guard let enableAutoSessionTracking = Bundle.main.object(
-            forInfoDictionaryKey:
-                FlutterSentryPlugin.SentryEnableAutoSessionTracking
-        ) else {
-            SentrySDK.start(options: [
-                "dsn": dsnString,
-            ])
-            return
+        let enableAutoSessionTracking = Bundle.main.object(
+            forInfoDictionaryKey: FlutterSentryPlugin.SentryEnableAutoSessionTracking)
+        // This is optional, skip if not set
+        if enableAutoSessionTracking != nil {
+            let value = enableAutoSessionTracking as? Bool
+            if value == nil {
+                NSException.raise(
+                    NSExceptionName.invalidArgumentException,
+                    format:"The value for key %@ is not a <boolean> type: %@",
+                    arguments: getVaList([
+                        Self.SentryEnableAutoSessionTracking,
+                        String(describing: value),
+                    ]))
+            }
+            options["enableAutoSessionTracking"] = value!
         }
 
-        guard let sessionTrackingIntervalMillis = Bundle.main.object(
-            forInfoDictionaryKey:
-                FlutterSentryPlugin.SentryEnableAutoSessionTracking
-        ) else {
-            SentrySDK.start(options: [
-                "dsn": dsnString,
-                "enableAutoSessionTracking": enableAutoSessionTracking,
-            ])
-            return
+        let sessionTrackingIntervalMillis = Bundle.main.object(
+            forInfoDictionaryKey: FlutterSentryPlugin.SentrySessionTrackingIntervalMillis)
+        // This is optional, skip if not set
+        if sessionTrackingIntervalMillis != nil {
+            let value = sessionTrackingIntervalMillis as? Int
+            if value == nil {
+                NSException.raise(
+                    NSExceptionName.invalidArgumentException,
+                    format:"The value for key %@ is not a <integer> type: %@",
+                    arguments: getVaList([
+                        Self.SentrySessionTrackingIntervalMillis,
+                        String(describing: value),
+                    ]))
+            }
+            options["sessionTrackingIntervalMillis"] = value!
         }
 
-        SentrySDK.start(options: [
-            "dsn": dsnString,
-            "enableAutoSessionTracking": enableAutoSessionTracking,
-            "sessionTrackingIntervalMillis": sessionTrackingIntervalMillis,
-        ])
+        SentrySDK.start(options: options)
     }
 
     public func handle(_ call: FlutterMethodCall,